LatticeAPI/lattice_api_client/requirements.txt
python-dotenv
requests
pytest
python-docx>=1.1